+++ AutoCareDiray.Shared/Models/Validation/RefillValidation.cs
@@ -16,7 +16,7 @@ namespace AutoCareDiray.Shared.Models.Validation
 
         public RefillValidation() { }
 
-        public bool ValidationMileage(int Mileage)
+        public bool ValidationMileage(double Mileage)
         {
             ErrorRemove(propertyMileage);
             if (Mileage >= 0)
